+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [3.0 Beta 5](https://github.com/sds100/KeyMapper/releases/tag/v3.0.0-beta.5)
 
-#### 4 April 2025
+#### 6 April 2025
 
 - #1625 HTTP Request action.